+++ src/util/namegen.cpp
@@ -44,7 +44,7 @@ string ProcNameGen::consonant()
 // Loads the namelists from the data files.
 void ProcNameGen::load_namelists()
 {
-    const uint8_t ftv_flags = FileX::FTV_FLAG_IGNORE_BLANK_LINES | FileX::FTV_FLAG_IGNORE_COMMENTS;
+    const unsigned int ftv_flags = FileX::FTV_FLAG_IGNORE_BLANK_LINES | FileX::FTV_FLAG_IGNORE_COMMENTS;
     names_f = FileX::file_to_vec(core().datafile("namegen/names-f.txt"), ftv_flags);
     names_m = FileX::file_to_vec(core().datafile("namegen/names-m.txt"), ftv_flags);
     names_s_a = FileX::file_to_vec(core().datafile("namegen/surname-a.txt"), ftv_flags);
